remove UnresolvedTableFunctionCall
